NetWorking using Python
Here i creat a socket which is a point of communication of a device. Then make
a connection to a host data.pr4e.org to retrieve a file called romeo.txt
so, i sent a Command to the host true the connection i creat using the socket
and the recieved data Stored in a cariable called data.